requirements.txt
jsonschema-2.6.0